commit before changing observers to function ptrs
Currently Observers are passed to Client as the whole object.
This will be another attempt at passing only the function ptrs,
which can be cloned and held by multiple other objects. This
will make it easier to create new connections as the observers
need not be passed from one Client to the next but can be instead
assigned to each new Client by the Factory.